+++ src/aac_metrics/__init__.py
@@ -26,6 +26,7 @@ from .classes.spice import SPICE
 from .classes.spider import SPIDEr
 from .classes.spider_fl import SPIDErFL
 from .classes.spider_max import SPIDErMax
+from .classes.vocab import Vocab
 from .functional.evaluate import evaluate, dcase2023_evaluate
 from .utils.paths import (
     get_default_cache_path,
@@ -52,6 +53,7 @@ __all__ = [
     "SPIDEr",
     "SPIDErFL",
     "SPIDErMax",
+    "Vocab",
     "evaluate",
     "dcase2023_evaluate",
     "get_default_cache_path",