+++ src/aac_metrics/functional/__init__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from .bleu import bleu
+from .bleu import bleu, bleu_1, bleu_2, bleu_3, bleu_4
 from .cider_d import cider_d
 from .evaluate import dcase2023_evaluate, evaluate
 from .fense import fense
@@ -13,10 +13,15 @@ from .spice import spice
 from .spider import spider
 from .spider_fl import spider_fl
 from .spider_max import spider_max
+from .vocab import vocab
 
 
 __all__ = [
     "bleu",
+    "bleu_1",
+    "bleu_2",
+    "bleu_3",
+    "bleu_4",
     "cider_d",
     "dcase2023_evaluate",
     "evaluate",
@@ -29,4 +34,5 @@ __all__ = [
     "spider",
     "spider_fl",
     "spider_max",
+    "vocab",
 ]