+++ tests/test_doc_examples.py
@@ -18,8 +18,9 @@ from aac_metrics.utils.tokenization import (
 
 class TestReadmeExamples(TestCase):
     def test_example_1(self) -> None:
-        if platform.system() == "Windows":
-            return None
+        # TODO : remove old check
+        # if platform.system() == "Windows":
+        #     return None
 
         candidates: list[str] = ["a man is speaking", "rain falls"]
         mult_references: list[list[str]] = [
@@ -58,8 +59,9 @@ class TestReadmeExamples(TestCase):
         )
 
     def test_example_2(self) -> None:
-        if platform.system() == "Windows":
-            return None
+        # TODO : remove old check
+        # if platform.system() == "Windows":
+        #     return None
 
         candidates: list[str] = ["a man is speaking", "rain falls"]
         mult_references: list[list[str]] = [