+++ CHANGELOG.md
@@ -11,8 +11,8 @@ All notable changes to this project will be documented in this file.
 - Rename metric `FluErr` to `FER`.
 
 ### Fixed
-- METEOR localization issue. ([#9](https://github.com/Labbeti/aac-metrics/issues/9))
-- SPIDErMax output when `return_all_scores=False`.
+- `METEOR` localization issue. ([#9](https://github.com/Labbeti/aac-metrics/issues/9))
+- `SPIDErMax` output when `return_all_scores=False`.
 
 ## [0.4.6] 2023-10-10
 ### Added